update MUI theme
