+++ packages/ui/src/ux.ts
@@ -1,11 +1,12 @@
-export * from './ux/Button';
-export * from './ux/AppWrapper';
 export * from './ux/AppHead';
-export * from './ux/TopMenu';
-export * from './ux/Page';
+export * from './ux/AppWrapper';
 export * from './ux/Breadcrumbs';
-export * from './ux/Section';
-export * from './ux/Typography';
+export * from './ux/Button';
 export * from './ux/Card';
 export * from './ux/Col';
 export * from './ux/ImgBox';
+export * from './ux/Page';
+export * from './ux/Section';
+export * from './ux/TextInput';
+export * from './ux/TopMenu';
+export * from './ux/Typography';
